+++ package.json
@@ -10,13 +10,17 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-navigation/bottom-tabs": "^6.5.8",
     "@react-navigation/native": "^6.1.7",
+    "@react-navigation/native-stack": "^6.9.13",
     "axios": "^1.4.0",
     "react": "18.2.0",
     "react-native": "0.71.8",
     "react-native-config": "^1.5.1",
     "react-native-geolocation-service": "^5.3.1",
     "react-native-get-location": "^3.0.1",
+    "react-native-safe-area-context": "^4.6.3",
+    "react-native-screens": "^3.22.0",
     "styled-components": "^5.3.1"
   },
   "devDependencies": {
