+++ pages/settings/Settings.tsx
@@ -1,0 +1,12 @@
+import { View, Text } from 'react-native'
+import React from 'react'
+
+const Settings = () => {
+  return (
+    <View>
+      <Text>Settings</Text>
+    </View>
+  )
+}
+
+export default Settings
